+++ src/main/java/ar/com/kfgodel/temas/application/auth/BackofficeCallbackAuthenticatorForRootsOnly.java
@@ -1,0 +1,29 @@
+package ar.com.kfgodel.temas.application.auth;
+
+import ar.com.kfgodel.dependencies.api.DependencyInjector;
+import ar.com.kfgodel.webbyconvention.api.auth.WebCredential;
+import convention.rest.api.tos.BackofficeUserTo;
+
+import java.util.Optional;
+
+/**
+ * Created by sandro on 03/07/17.
+ */
+public class BackofficeCallbackAuthenticatorForRootsOnly extends BackofficeCallbackAuthenticator {
+
+    @Override
+    public Optional<Object> apply(WebCredential webCredential) {
+        Optional<BackofficeUserTo> backofficeUser = createToFrom(webCredential);
+        Optional<Long> idAutenticado = backofficeUser
+                .filter(BackofficeUserTo::hasGrantedAccess)
+                .filter(BackofficeUserTo::isRoot)
+                .map(this::buscarIdDelUsuarioEnLaBase);
+        return idAutenticado.map(Object.class::cast); // El casteo solo para ayudar al compilador
+    }
+
+    public static BackofficeCallbackAuthenticator create(DependencyInjector appInjector) {
+        BackofficeCallbackAuthenticator authenticator = new BackofficeCallbackAuthenticatorForRootsOnly();
+        authenticator.appInjector = appInjector;
+        return authenticator;
+    }
+}
